wip: outline CodeWriter refactor
